add docs, fix overloading, deduplication of tests, removed integration testing not using databricks sdk
