+++ docs/dqx/docs/guide/quality_checks_definition.mdx
@@ -720,6 +720,60 @@ Example checks saved in a Delta or Lakebase table (compact format — `for_each_
 
 If `run_config_name` is not provided, "default" is used. Typically, the input table or job name is used for run config name to establish a one-to-one mapping between tables or jobs and checks.
 
+## Variable Substitution
+
+DQX supports variable substitution in declarative check definitions (YAML, JSON, or Delta tables). This allows you to parameterize your quality rules and inject values at **load time** via the `variables` parameter in `load_checks`.
+
+### Syntax and Scope
+
+Placeholders are defined using the `{{ variable_name }}` syntax. Variable substitution is supported in **all string values** within the check definitions, including:
+- `name`
+- `filter`
+- `check` function arguments (`arguments`) and column names (`for_each_column`)
+- any other top-level or nested string field
+
+### Resolution
+
+Variables are resolved at **load time** when the checks are loaded from the storage backend. To resolve variables, pass a dictionary to the `variables` parameter of the `load_checks` method.
+
+<Admonition type="info" title="Note">
+Variable substitution is only available when defining checks declaratively (as dictionaries or in files/tables). It is not supported when using DQX classes (e.g., `DQRowRule`) directly.
+</Admonition>
+
+<Tabs>
+  <TabItem value="YAML" label="YAML" default>
+    ```yaml
+    - criticality: error
+      check:
+        function: is_in_range
+        arguments:
+          column: temperature
+          min_limit: {{ min_temp }}
+          max_limit: {{ max_temp }}
+      filter: "region = '{{ region }}'"
+    ```
+  </TabItem>
+  <TabItem value="Python" label="Python">
+    ```python
+    from databricks.labs.dqx.engine import DQEngine
+    from databricks.labs.dqx.config import FileChecksStorageConfig
+    from databricks.sdk import WorkspaceClient
+
+    dq_engine = DQEngine(WorkspaceClient())
+
+    # Load checks with variable resolution
+    resolved_checks = dq_engine.load_checks(
+        config=FileChecksStorageConfig(location="checks.yml"),
+        variables={
+            "min_temp": 0,
+            "max_temp": 100,
+            "region": "EMEA"
+        }
+    )
+    ```
+  </TabItem>
+</Tabs>
+
 ## Validating syntax of quality checks
 
 You can validate the syntax of checks loaded from a storage system or checks defined programmatically before applying them.
