+++ docs/dqx/docs/reference/quality_checks.mdx
@@ -18,6 +18,10 @@ All rule types, including row-level and dataset-level rules, can be defined and 
 
 You can explore the implementation details of the check functions [here](https://github.com/databrickslabs/dqx/blob/v0.13.0/src/databricks/labs/dqx/check_funcs.py).
 
+<Admonition type="tip" title="Variable Substitution">
+All declarative check definitions (YAML, JSON, or Delta tables) support **variable substitution** for string-based fields using the `{{ variable_name }}` syntax. This allows for dynamic parameterization of column names, thresholds, and filters at load time. See the [User Guide](/docs/guide/quality_checks_definition/#variable-substitution) for more details.
+</Admonition>
+
 ## Row-level checks reference
 
 Row-level checks are applied to each row in a PySpark DataFrame. The quality check results are reported for individual rows in the result columns.
